+++ example/lib/main.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'dart:async';
-import 'dart:typed_data';
 import 'dart:ui' as ui;
 import 'dart:math' show Point;
 
-import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:rectangle_detector/rectangle_detector.dart';
 import 'package:image_picker/image_picker.dart';
@@ -36,7 +34,6 @@ class RectangleDetectorDemo extends StatefulWidget {
 }
 
 class _RectangleDetectorDemoState extends State<RectangleDetectorDemo> {
-  final _rectangleDetector = RectangleDetector();
   List<RectangleFeature> _detectedRectangles = [];
   bool _isDetecting = false;
   ui.Image? _image;
@@ -140,7 +137,7 @@ class _RectangleDetectorDemoState extends State<RectangleDetectorDemo> {
       final Uint8List imageBytes = byteData.buffer.asUint8List();
       
       // 调用插件检测矩形
-       final List<RectangleFeature> rectangles = await _rectangleDetector.detectAllRectangles(imageBytes);
+       final List<RectangleFeature> rectangles = await RectangleDetector.detectAllRectangles(imageBytes);
       
       if (rectangles.isNotEmpty) {
         setState(() {